Add Effects-Components
